backend/sample_jwt/sample_jwt/urls.py
from django.contrib import admin
from django.conf import settings
from django.urls import path, include, re_path
from django.conf.urls.static import static

# Swagger imports
from rest_framework import permissions
from drf_yasg.views import get_schema_view
from drf_yasg import openapi

# Swagger configuration
schema_view = get_schema_view(
    openapi.Info(
        title="JWT API",
        default_version='v1',
        description="API documentation for JWT Web Application",
    ),
    public=True,
    permission_classes=[permissions.AllowAny],
)

urlpatterns = [
    
    path('', include('users.urls')),

    # ✅ Swagger endpoints
    re_path(r'^swagger/$', schema_view.with_ui('swagger', cache_timeout=0)),
    re_path(r'^redoc/$', schema_view.with_ui('redoc', cache_timeout=0)),
]

# Media files
if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)