backend/sample_jwt/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-paypal==2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fastapi==0.116.1
fonttools==4.56.0
git-filter-repo==2.47.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-dialogflow==2.39.1
googleapis-common-protos==1.69.0
greenlet==3.2.3
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
idna==3.10
lxml==5.3.0
oauthlib==3.2.2
paypalrestsdk==1.13.3
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyphen==0.17.2
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
razorpay==1.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
social-auth-core==4.5.4
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.35.0
weasyprint==64.1
webencodings==0.5.1
xlwt==1.3.0
zopfli==0.2.3.post1
